Allow Client-Specified Secret Overlap Duration
I'm _very tempted_ to try to limit this to a day or less, but I don't
think I can do that in a regex without tying myself in knots. I'll at
least make sure it starts with (an optional plus sign followed by) a P
so that there's a hope of validity. I don't expect anyone will adjust
this value, frankly, but there's little harm in the possibility.